authData has no consumers anymore.
To stop computing content of authData here, was the goal of this patch series.

It stored credentials only for the last proxy configuration with the same
proxy address and "forgot" earlier configurations.
Disabled proxies were not sorted out, different ports were ignored

In case of different proxy configurations with the same address, always the
same username and password (from the last configuration) was used.